handle split response for already-split blocks differently
It's possible (probable, even!) that splitting a block will
result in a non-last block that still has >1/2 size entries
in it. If this happens and we try to write again, but there
is already a hardtail entry, this isn't an error case --
change the logic to detect a hardtail already present and
return Ok, rather than reporting No_space, if the split
recommendation isn't Split_emergency (which happens when
the requested info *can't fit* in the block, rather than
when a split is *desirable* but not necessary).