refactor blocks to store their hardtails out of the commit list
This fixes a major bug with deletions. Previously they would fail
silently when the entries for the ID corresponding to the requested
deletion weren't in the first block of the directory.

There remains work to do around correct handling of entries when
deserializing blocks (currently we both put the hardtail in the commit
list, and encode it into the block record), but this at least fixes
removals most of the time, and gives us a fix for valid_after to boot.