Break cycle between Dir and Entry (dune 3.7)
There is a dependency cycle in `lib`: `entry.ml` depends on `Dir`, and
`dir.mli` depends on `Entry`.

Dune used to be OK with that, but the new dependency analysis algorithm
in 3.7 makes tighter checks and (correctly) detects it. It's kinda OK in
that particular case because `Dir` only uses a type alias from `Entry`,
but dune can not reliably detect this.

So this PR adds an extra `Entry0` module with just the type definition,
in a way that splits the cycle and makes it compatible with dune 3.7.